FlatAdmin yüklendi. dal ve bll tamamlandı.

+++ Project.COREMVC/Models/RecipeDetails/RequestModels/RecipeDetailRequestModel.cs
@@ -5,7 +5,7 @@
         public int RecipeID { get; set; }
         public int IngredientID { get; set; }
         public string Instruction { get; set; }
-        public decimal IngredientQuantity { get; set; }
+        public int IngredientQuantity { get; set; }
         public string Unit { get; set; }
 
     }
